requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
backoff==2.2.1
click==8.1.7
Deprecated==1.2.14
fastapi==0.109.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
protobuf==4.25.2
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
wrapt==1.16.0
zipp==3.17.0
